# perf: Only run the datascript-devtools in dev mode