I noticed that the listener was being run in prod as well as debug mode, so I added a conditional to have it run only in debug mode. 

I expect this will create significant speedup throughout the entire app (anywhere that the listener is registered to listen to). One place I measured it was when you create a new block by hitting "ENTER".

- Before this change, it took ~400ms (`~1900ms - ~1500ms = 400ms`):

  ![image](https://user-images.githubusercontent.com/6979755/119247584-b7cb4400-bb58-11eb-83a7-825005cbdb3e.png)


- Now, it takes ~200ms (`~3500ms - ~3300ms = 400ms`):

   ![image](https://user-images.githubusercontent.com/6979755/119247546-658a2300-bb58-11eb-9ef9-7c8afed7b253.png)


<br/>


<details><summary>If you want to view the difference yourself, here are the steps to replicate the above...</summary>

<br/>

**First, test the speed in dev mode (i.e. `config/debug? == true`):**
1. Pull down the change. Create a block with random content in it.
2. Start recording with the Chrome flamegraph (under the Performance tab of devtools).
3. Click into the block you created in step 1, hit ENTER to create a new block, then stop the flamegraph recording.

**Now, test the speed when you invert `config/debug?`:**
1. Temporarily wrap the `config/debug?` with a `not` (i.e. make line 401 `(when (not config/debug?)`).
2. Hard refresh the app. This is important! A normal hot-reload won't get rid of the listener because it gets registered at the top level of the app.
3. Create a block with random content in it.
4. Start recording with the Chrome flamegraph (under the Performance tab of devtools).
5. Click into the block you created in step 1, hit ENTER to create a new block, then stop the flamegraph recording.

</details>

---

P.S. I read through the [contributing guidelines](https://athensresearch.gitbook.io/handbook/building-athens-build-develop-contribute/contributing), but this is my first PR so there's a high chance I didn't do something quite right. Apologies if so! I'm very open to feedback if I missed something.